+++ waltid-libraries/protocols/waltid-openid4vci/src/commonMain/kotlin/id/walt/openid4vci/repository/authorization/AuthorizationCodeRepository.kt
@@ -1,9 +1,5 @@
 package id.walt.openid4vci.repository.authorization
 
-import id.walt.openid4vci.Session
-import kotlinx.serialization.Serializable
-import kotlin.time.Instant
-
 /**
  * Storage abstraction for authorization code sessions.
  */
@@ -14,27 +10,3 @@ interface AuthorizationCodeRepository {
 }
 
 class DuplicateCodeException : IllegalStateException("Code collision detected")
-
-/**
- * Authorization code payload. Applications can supply custom implementations if they need extra fields.
- */
-interface AuthorizationCodeRecord {
-    val code: String
-    val clientId: String
-    val redirectUri: String?
-    val grantedScopes: Set<String>
-    val grantedAudience: Set<String>
-    val session: Session
-    val expiresAt: Instant
-}
-
-@Serializable
-data class DefaultAuthorizationCodeRecord(
-    override val code: String,
-    override val clientId: String,
-    override val redirectUri: String?,
-    override val grantedScopes: Set<String>,
-    override val grantedAudience: Set<String>,
-    override val session: Session,
-    override val expiresAt: Instant,
-) : AuthorizationCodeRecord
